Split tokenizer into separate file and merge character validation into it
